Skip adding Ruby + Python to PATH if they don't exist
Append (N) to these globs which applies the nullglob flag, i.e. they
will expand to empty string if there are no matching files. The default
is to print an error. See https://stackoverflow.com/questions/71729812/is-there-a-zsh-equivalent-to-shopt-s-nullglob

Note that this will create a blank entry in PATH if Ruby is not
installed (i.e. the string :: will appear somewhere in PATH).
Empirically this seems to not cause problems.